resolved_ts,test_sst_importer: simpify test code
