+++ components/resolved_ts/tests/integrations/mod.rs
@@ -69,10 +69,13 @@ fn test_resolved_ts_basic() {
     meta.set_region_id(r1.id);
     meta.set_region_epoch(sst_epoch);
 
-    suite.upload_sst(r1.id, &meta, &data).unwrap();
+    let import = suite.get_import_client(r1.id);
+    send_upload_sst(import, &meta, &data).unwrap();
 
     let tracked_index_before = suite.region_tracked_index(r1.id);
-    suite.must_ingest_sst(r1.id, meta);
+    let ctx = suite.get_context(r1.id);
+    let import = suite.get_import_client(r1.id);
+    must_ingest_sst(import, ctx, meta);
     let mut tracked_index_after = suite.region_tracked_index(r1.id);
     for _ in 0..10 {
         if tracked_index_after > tracked_index_before {
